+++ main.py
@@ -1,4 +1,22 @@
-from call import os, clear, settings, name, loadpath, autoupdt, firstrun, readchar, audiod, videod, subd, vidhevc, update, slpath, about, debug
+import os
+try:
+    from getch import getch #py-getch
+    from colorama import init, Fore, Back, Style
+except ModuleNotFoundError:
+    print("You are missing dependencies do you want to install them with pip? [Y/n]")
+    cmd = input("#")
+    if(cmd == "y" or cmd == "Y"):
+        os.system("pip install -r requirements.txt")
+        print("\n\nif all went well, restart yt-dl and it will work\nIf not run \"pip install -r requirements.txt\"\nYou should also install ffmpeg and add it to PATH")
+        input("press any key to quit")
+        exit()
+    else:
+        print("install them before you can use yt-dl\nYou should also install ffmpeg and add it to PATH")
+        input("press any key to quit")
+        exit()
+
+from kolreq import clear, readchar
+from call import settings, name, loadpath, autoupdt, firstrun, audiod, videod, subd, vidhevc, update, slpath, about, debug
 
 clear()