+++ main.py
@@ -1,4 +1,29 @@
-from call import os, clear, settings, name, loadpath, autoupdt, firstrun, readchar, audiod, videod, subd, vidhevc, update, slpath, about, debug
+import os
+try:
+    from getch import getch #py-getch
+    from colorama import init, Fore, Back, Style
+except ModuleNotFoundError:
+    print("You are missing dependencies do you want to install them with pip? [Y/n]")
+    cmd = input("#")
+    if(cmd == "y" or cmd == "Y"):
+        print("Do you with to update pip first? If yes type the name of you python exectable.")
+        cmd = input("#")
+        if (cmd != ""):
+            py = cmd
+            print("updating pip...")
+            os.system(f"{py} -m pip install -U pip")
+        print("installing dependecies...")
+        os.system("pip install -r requirements.txt")
+        print("\n\nif all went well, restart yt-dl and it will work\nIf not run \"pip install -r requirements.txt\"\nYou should also install ffmpeg and add it to PATH")
+        input("press any key to quit")
+        exit()
+    else:
+        print("install them before you can use yt-dl\nYou should also install ffmpeg and add it to PATH")
+        input("press any key to quit")
+        exit()
+
+from kolreq import clear, readchar
+from call import settings, name, loadpath, autoupdt, firstrun, audiod, videod, subd, vidhevc, update, slpath, about, debug
 
 clear()
 
@@ -10,7 +35,7 @@ else:
     firstrun()
 
 while(True):
-    print("1. Audio\t5. Vid to Hevc\n2. Video\t6. Update\n3. Exit\t\t7. Download path\n4. Subtitles\t8. About, Changelog")
+    print("1. Audio\t5. Re-encode\n2. Video\t6. Update\n3. Exit\t\t7. Download path\n4. Subtitles\t8. About, Changelog")
     cmd = readchar("#")
 
     if(cmd == "1"): #audio