+++ whatsnew.md
@@ -28,7 +28,7 @@ Osana is still not finished, lol.
 To do by severity 2.1.6:
         [    ] (high)    Remake firstrun and integrate dependenci install
         [    ] (high)    split some part of call.py into kolreq.py  
-I'wont  [    ] (low)     PATH export ffmpeg to the yt-dl dir.
+I'wont  [    ] (low)     PATH export ffmpeg to the yt-dl dir. is ffmpeg in yt-dl dir
 I'wont  [    ] (low)     Ctrl+D to quit
 
 
@@ -38,7 +38,7 @@ how to git:
 testing:
 git add . (stage changes)
 git commit -m "message" (commit)
-git push (push)
+git push (push)NO
 
 master:
 git checkout master (chage to master)