+++ whatsnew.md
@@ -1,8 +1,16 @@
+New in 2.1.6
+First run, dependecy update fix
+added support link to my discord https://discord.gg/W88375j
+added NVENC h264 encoding 
+quality of life improvements
+
 New in 2.1.5
 Updating pip before updating dependencies
 autoupdate, upgrading dependencies
 convert any video to HEVC
 generate powershell launch script
+got rid of wildcard import
+got rid of unused imports
 if you are missing dependencies we will help you get them
 
 New in 2.1.4
@@ -13,5 +21,12 @@ added a text so the program does not look like it froze when it's starting youtu
 dependecy upglade uses the pip name you set
 launch script generator
 working autoupdate with a toggle
+
+New in 2.1.3
+new versioning system, Lang, Featureset, patch/bugfix
+Adding new settings will nolonger break on first launch.
+Python & pip executable name is now saved in settings
+fixed very stupid bugs. I mean 0. Goback now works on Videos & subs.
+new debug feature
 
 Osana is still not finished, lol.